requirements.txt
selenium
chromedriver_autoinstaller
instaloader
requests
python-dotenv
pytesseract
pyautogui
keyboard
pillow
opencv-python